input items add inputPanel.go
